+++ src/bewegung/core/vector/array3d.py
@@ -115,9 +115,9 @@ class VectorArray3D(VectorArray3DABC):
 
         return np.allclose(self.x, other.x) and np.allclose(self.y, other.y) and np.allclose(self.z, other.z)
 
-    def __add__(self, other: VectorArray3DABC) -> VectorArray3DABC:
+    def __add__(self, other: Union[VectorArray3DABC, Vector3D]) -> VectorArray3DABC:
         """
-        Add operation between vector arrays
+        Add operation between vector arrays or a vector array and a vector
 
         Args:
             other : Another vector array of equal length
@@ -127,9 +127,13 @@ class VectorArray3D(VectorArray3DABC):
         assert self.dtype == other.dtype
         return VectorArray3D(self.x + other.x, self.y + other.y, self.z + other.z)
 
-    def __sub__(self, other: VectorArray3DABC) -> VectorArray3DABC:
+    def __radd__(self, *args, **kwargs):
+
+        return self.__add__(*args, **kwargs)
+
+    def __sub__(self, other: Union[VectorArray3DABC, Vector3D]) -> VectorArray3DABC:
         """
-        Substract operator between vector arrays
+        Substract operator between vector arrays or a vector array and a vector
 
         Args:
             other : Another vector array of equal length
@@ -138,6 +142,10 @@ class VectorArray3D(VectorArray3DABC):
         assert len(self) == len(other)
         assert self.dtype == other.dtype
         return VectorArray3D(self.x - other.x, self.y - other.y, self.z - other.z)
+
+    def __rsub__(self, *args, **kwargs):
+
+        return self.__sub__(*args, **kwargs)
 
     def __mul__(self, other: Number) -> VectorArray3DABC:
         """